Moved low-level utilities for the app into separate file app_utils.py Added export_script.py file with code to export scripts.
